genrate/requirements.txt
PyPDF2==3.0.1
pdfplumber==0.10.3
ollama==0.1.7
Pillow==10.1.0
pyttsx3==2.90
opencv-python-headless==4.8.1.78
numpy==1.24.3
requests==2.31.0 